petites amélios

+++ src/app/features/exercices/exo06/exo06.ts
@@ -18,18 +18,38 @@ export class Exo06 implements OnInit {
   private _LibraryService = inject(LibraryService);
 
   allBooks: Book[] = [];
+  filteredBooks: Book[] = [];
+
 
   ngOnInit() {
+    this.refreshList();
+  }
+
+  refreshList() {
     this.allBooks = this._LibraryService.getBooks();
+    this.filteredBooks = [...this.allBooks];
   }
 
   onAddBook(newBook: Book) {
     this._LibraryService.addBook(newBook);
-    this.allBooks = this._LibraryService.getBooks();
+    this.refreshList()
   }
 
   onDeleteBook(index: number) {
-    this._LibraryService.removeBook(index);
-    this.allBooks = this._LibraryService.getBooks();
+
+    const bookTitle = this.allBooks[index].title;
+
+    if (confirm(`Voulez vous vraiment supprimer le livre ${bookTitle}`)) {
+      this._LibraryService.removeBook(index);
+      this.refreshList();
+    }
+  }
+
+  onSearch(event: Event) {
+    const search = (event.target as HTMLInputElement).value.toLowerCase();
+
+    this.filteredBooks = this.allBooks.filter(b =>
+    b.title.toLowerCase().includes(search) || b.author.toLowerCase().includes(search)
+    );
   }
 }
